Test API key when load admin screen and edit key.
This improved UX because otherwise the only way to test if everything is working is to write a rude comment on the site. This also helps surface problems that come up unexpectedly (e.g. the API key is over its quota), because the front-end has to fail gracefully and let comments through.