Restructure Admin to make room for an Impact tab.
Previously the only things on the page were settings, but now we want to have stats as well. This creates tabbed navigation where Settings and Impact each have their own tab. This required refactoring all of the settings logic/UI into it's own component. The tabs use `pushState()` to make deep-linking possible, and avoid "breaking the web."

Additionally, `savedSettings` is now reset when settings are changed after saving. That was included in this commit because it would have been difficult to separate it out into an atomic one.

This doesn't add any of the stats functionality, that will come in a later commit. This just sets up the infrastructure necessary to have multiple tabbed screens on the page.